+++ app/src/main/java/care/intouch/app/feature/diary/data/EmotionsRepository.kt
@@ -1,8 +1,8 @@
 package care.intouch.app.feature.diary.data
 
-import care.intouch.app.feature.diary.presentation.ui.EmotionScreens.models.EmotionDesc
-import care.intouch.app.feature.diary.presentation.ui.EmotionScreens.models.EmotionDescriptionEnum
-import care.intouch.app.feature.diary.presentation.ui.EmotionScreens.models.EmotionDescriptionTask
+import care.intouch.app.feature.diary.presentation.ui.emotionScreen.models.EmotionDesc
+import care.intouch.app.feature.diary.presentation.ui.emotionScreen.models.EmotionDescriptionEnum
+import care.intouch.app.feature.diary.presentation.ui.emotionScreen.models.EmotionDescriptionTask
 
 interface EmotionsRepository {
     suspend fun saveEmotion(emotionDesc: EmotionDesc)
