+++ app/src/main/java/care/intouch/app/feature/diary/domain/useCase/SaveEmotionUC.kt
@@ -1,7 +1,7 @@
 package care.intouch.app.feature.diary.domain.useCase
 
 import care.intouch.app.feature.diary.data.EmotionsRepository
-import care.intouch.app.feature.diary.presentation.ui.EmotionScreens.models.EmotionDesc
+import care.intouch.app.feature.diary.presentation.ui.emotionScreen.models.EmotionDesc
 import javax.inject.Inject
 
 interface SaveEmotionUC {
