--- app/src/main/java/care/intouch/app/feature/diary/presentation/ui/EmotionScreens/models/EmotionDesc.kt
@@ -1,3 +1,0 @@
-package care.intouch.app.feature.diary.presentation.ui.EmotionScreens.models
-
-enum class EmotionDesc { TERRIBLE, BAD, OKAY, GOOD, GREAT, EMPTY }
