+++ app/src/main/java/care/intouch/app/feature/diary/presentation/ui/emotionScreen/models/EmotionDesc.kt
@@ -1,0 +1,3 @@
+package care.intouch.app.feature.diary.presentation.ui.emotionScreen.models
+
+enum class EmotionDesc { TERRIBLE, BAD, OKAY, GOOD, GREAT, EMPTY }
