+++ app/src/main/java/care/intouch/app/feature/diary/presentation/ui/fillingOutScreen/models/FillingOutScreenState.kt
@@ -2,8 +2,8 @@ package care.intouch.app.feature.diary.presentation.ui.fillingOutScreen.models
 
 
 data class FillingOutScreenState(
-    val detailsText: String,
-    val analysisText: String,
-    val typeText: String,
-    val sensationsText: String,
+    val detailsText: String = "",
+    val analysisText: String = "",
+    val typeText: String = "",
+    val sensationsText: String = "",
 )
