+++ app/src/main/java/care/intouch/app/feature/diary/presentation/ui/fillingOutScreen/viewModel/FillingOutViewModel.kt
@@ -7,12 +7,11 @@ import care.intouch.app.feature.diary.domain.useCase.GetEmotionDescUC
 import care.intouch.app.feature.diary.domain.useCase.GetEmotionUC
 import care.intouch.app.feature.diary.domain.useCase.GetSavedAnswers
 import care.intouch.app.feature.diary.domain.useCase.SaveAnswersUC
-import care.intouch.app.feature.diary.presentation.ui.EmotionScreens.models.EmotionDescriptionEnum
-import care.intouch.app.feature.diary.presentation.ui.EmotionScreens.models.EmotionDescriptionTask
+import care.intouch.app.feature.diary.presentation.ui.emotionScreen.models.EmotionDescriptionTask
+import care.intouch.app.feature.diary.presentation.ui.fillingOutScreen.getEmotionDesc
 import care.intouch.app.feature.diary.presentation.ui.fillingOutScreen.models.FillingOutDataEvent
 import care.intouch.app.feature.diary.presentation.ui.fillingOutScreen.models.FillingOutScreenState
 import care.intouch.uikit.common.ImageVO
-import care.intouch.uikit.common.StringVO
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -30,14 +29,7 @@ class FillingOutViewModel @Inject constructor(
     private val saveAnswersUC: SaveAnswersUC,
     private val getSavedAnswers: GetSavedAnswers,
 ) : ViewModel() {
-    private val _uiState = MutableStateFlow(
-        FillingOutScreenState(
-            detailsText = "",
-            analysisText = "",
-            typeText = "",
-            sensationsText = "",
-        )
-    )
+    private val _uiState = MutableStateFlow(FillingOutScreenState())
     val uiState = _uiState.asStateFlow()
     private val _sharedImageUiState = MutableSharedFlow<ImageVO?>()
     val sharedImageUiState = _sharedImageUiState.asSharedFlow()
@@ -136,266 +128,7 @@ class FillingOutViewModel @Inject constructor(
     private fun getEmotionDescription() {
         val result: MutableList<EmotionDescriptionTask> = mutableListOf()
         viewModelScope.launch {
-            getEmotionDescUC.invoke().forEach {
-                when (it) {
-                    EmotionDescriptionEnum.Loss -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(
-                                care.intouch.app.R.string.loss_clarifying_emotional
-                            ), EmotionDescriptionEnum.Loss
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Humility -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.humility_clarifying_emotional),
-                            EmotionDescriptionEnum.Humility
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Anxiety -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(
-                                care.intouch.app.R.string.anxiety_clarifying_emotional
-                            ), EmotionDescriptionEnum.Anxiety
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Embarrassment -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.embarrassment_clarifying_emotional),
-                            EmotionDescriptionEnum.Embarrassment
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Depression -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.depression_clarifying_emotional),
-                            EmotionDescriptionEnum.Depression
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Fear -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(
-                                care.intouch.app.R.string.fear_clarifying_emotional
-                            ), EmotionDescriptionEnum.Fear
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Anger -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.anger_clarifying_emotional),
-                            EmotionDescriptionEnum.Anger
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Excitement -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(
-                                care.intouch.app.R.string.excitement_clarifying_emotional
-                            ), EmotionDescriptionEnum.Excitement
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Shame -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.shame_clarifying_emotional),
-                            EmotionDescriptionEnum.Shame
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Disgust -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.disgust_clarifying_emotional),
-                            EmotionDescriptionEnum.Disgust
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Guilt -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.guilt_clarifying_emotional),
-                            EmotionDescriptionEnum.Guilt
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Nervousness -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.nervousness_clarifying_emotional),
-                            EmotionDescriptionEnum.Nervousness
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Sadness -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.sadness_clarifying_emotional),
-                            EmotionDescriptionEnum.Sadness
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Exhaustion -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.exhaustion_clarifying_emotional),
-                            EmotionDescriptionEnum.Exhaustion
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Impatience -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.impatience_clarifying_emotional),
-                            EmotionDescriptionEnum.Impatience
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Disappointment -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.disappointment_clarifying_emotional),
-                            EmotionDescriptionEnum.Disappointment
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Confusion -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.confusion_clarifying_emotional),
-                            EmotionDescriptionEnum.Confusion
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Loneliness -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.loneliness_clarifying_emotional),
-                            EmotionDescriptionEnum.Loneliness
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Acceptance -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.acceptance_clarifying_emotional),
-                            EmotionDescriptionEnum.Acceptance
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Frustration -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.frustration_clarifying_emotional),
-                            EmotionDescriptionEnum.Frustration
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Laziness -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.laziness_clarifying_emotional),
-                            EmotionDescriptionEnum.Laziness
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Pride -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.pride_clarifying_emotional),
-                            EmotionDescriptionEnum.Pride
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Hope -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.hope_clarifying_emotional),
-                            EmotionDescriptionEnum.Hope
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Calmness -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.calmness_clarifying_emotional),
-                            EmotionDescriptionEnum.Calmness
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Gratitude -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.gratitude_clarifying_emotional),
-                            EmotionDescriptionEnum.Gratitude
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Interest -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.interest_clarifying_emotional),
-                            EmotionDescriptionEnum.Interest
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Respect -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.respect_clarifying_emotional),
-                            EmotionDescriptionEnum.Respect
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Enthusiasm -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.enthusiasm_clarifying_emotional),
-                            EmotionDescriptionEnum.Enthusiasm
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Happiness -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.happiness_clarifying_emotional),
-                            EmotionDescriptionEnum.Happiness
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Inspiration -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.inspiration_clarifying_emotional),
-                            EmotionDescriptionEnum.Inspiration
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Joy -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(
-                                care.intouch.app.R.string.joy_clarifying_emotional
-                            ), EmotionDescriptionEnum.Joy
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Love -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.love_clarifying_emotional),
-                            EmotionDescriptionEnum.Love
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Amazement -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.amazement_clarifying_emotional),
-                            EmotionDescriptionEnum.Amazement
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Satisfaction -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.satisfaction_clarifying_emotional),
-                            EmotionDescriptionEnum.Satisfaction
-                        )
-                    )
-
-                    EmotionDescriptionEnum.SelfLove -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.self_love_clarifying_emotional),
-                            EmotionDescriptionEnum.SelfLove
-                        )
-                    )
-
-                    EmotionDescriptionEnum.Euphoria -> result.add(
-                        EmotionDescriptionTask(
-                            StringVO.Resource(care.intouch.app.R.string.euphoria_clarifying_emotional),
-                            EmotionDescriptionEnum.Euphoria
-                        )
-                    )
-                }
-            }
+            result.addAll(getEmotionDesc(getEmotionDescUC.invoke()))
             _sharedListUiState.emit(result)
         }
     }
